\name{flowVS-package}
\alias{flowVS-package}
\alias{flowVS}
\docType{package}
\title{
 flowVS: Variance stabilization in flow cytometry (and microarrays).
}

\description{
Please see the vignette.
}


\author{
  Ariful Azad <azad@lbl.gov>
}


\references{
  Ariful Azad, Bartek Rajwa, and Alex Pothen (2015), "flowVS: Channel-Specific Variance Stabilization in Flow Cytometry", manuscript submitted for publication.
}


\seealso{
  \code{\link{transFlowVS}}, \code{\link{microVS}}
}

\keyword{ transformation }
\keyword{ variance stabilization }
